code: Fix errors hidden by parent scope leaking into function
The variable scope in CMake is unlike other languages, and simply prevents functions from unintentionally polluting the parent scope. Unfortunately this also means that we need to take extra measurements to guarantee a good default state, and that 'DEFINED' is a useless keyword. Therefore by using a random prefix, we can prevent unintentional pollution of the function scope by the caller.